+++ Implicit-Lagrange/Topology_Optimization/Strain_Energy_Minimize.h
@@ -107,7 +107,7 @@ private:
 
 public:
   bool nodal_density_flag_;
-  size_t last_comm_step, last_solve_step, current_step;
+  int last_comm_step, last_solve_step, current_step;
   std::string my_fea_module = "Elasticity";
 
   StrainEnergyMinimize_TopOpt(FEA_Module *FEM, bool nodal_density_flag)
